Fix: add correct constraints and types on migrations

+++ src/DataAccess/migrations/2024_12_20_123327_create_universes_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('description')->nullable();
             $table->boolean('is_archived')->default(false);
-            $table->integer("creator_target_id")->comment('Gets the actual Id of the creator of the universe. Group Id if it is a group, user Id if it is a user');
-            $table->bigInteger('root_place_id')->nullable();
+            $table->unsignedBigInteger("creator_target_id")->comment('Gets the actual Id of the creator of the universe. Group Id if it is a group, user Id if it is a user');
+            $table->unsignedBigInteger('root_place_id')->nullable();
             $table->string("privacy_type");
             $table->smallInteger("creator_type");
             $table->boolean('api_services')->nullable()->comment('Allows for datastores, httpservice and other funny studio things');
